+++ wdio.conf.js
@@ -1,7 +1,8 @@
 const binaryPath = process.env.BINARY || 'C:\\Users\\Nguyen Luan\\AppData\\Local\\Programs\\msgr-client-electron\\HalomeSB.exe';
 
 exports.config = {
-    specs: ['./test/specs/**/*.ts'],
+    // specs: ['./test/specs/**/*.ts'],
+    specs: ['./test/specs/**/hahalolo.e2e.ts'],
     exclude: [],
     maxInstances: 10,
     capabilities: [
